Make window size into a parameter
This required a refactoring and cleanup of the ratio calculation, which
incidentally makes the code a lot more readable.

At the same time, instead of saving the whole data vector, I am now only
storing a summary statistic and the description of the ECDF. This will
make it possible to run and plot several window sizes without taxing the
memory too much.